requirement_new.txt
Brotli==1.0.9
certifi==2024.2.2
charset-normalizer==2.0.4
contourpy==1.1.1
cupy==12.3.0
cycler==0.12.1
easydict==1.13
fastrlock==0.8.2
filelock==3.13.1
fonttools==4.51.0
gmpy2==2.1.2
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.3
matmul==0.0.0
matplotlib==3.7.5
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pip==24.0
protobuf==5.26.1
pyinn==0.0.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
scipy==1.10.1
setuptools==69.5.1
six==1.16.0
sympy==1.12
tensorboardX==2.6.2.2
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
triton==2.2.0
typing_extensions==4.11.0
urllib3==2.2.1
wheel==0.43.0
zipp==3.18.1


#argparse
#easydict
#json
#numpy
#matplotlib
#scipy
#opencv-python
#pprint
#torchvision
#tensorboardX

